\name{aggreg.judgement-data}
\docType{data}
\alias{aggreg.judgement}
\alias{aggregJudgement}
\alias{aggreg.judgement-data}
\title{A pairwise comparison matrix (PCM)}
\description{
Aggreg.judgement includes a PCM computed from aggregation of decision makers opinions. The PCM includes 4 criteria to prioritize a set of genes of interest according to their potential to serve as a tumor marker/therapeutic target. The criteria include: Tumor_expression, Normal_expression, Frequency, and Epitopes.
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
E. Forman and K. Peniwati. Aggregating individual judgments and priorities with the analytic hierarchy process. European Journal of Operational Research, 1998.
}


