% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahmatrix.R
\name{ahmatrix}
\alias{ahmatrix}
\title{ahmatrix}
\usage{
ahmatrix(x)
}
\arguments{
\item{x}{a numeric triangular matrix, where empty elements are specified with NA.}
}
\value{
An S4 object including a PCM.
}
\description{
Converting a triangular matrix into a square pairwise comparison matrix (PCM) where the diagonal values are equal to 1 and a[i,j] = 1/a[j,i].
}
\examples{
mat <- matrix(nrow = 3, ncol = 3, data = NA)
mat[1,2] <- 5
mat[1,3] <- 2
mat[2,3] <- 7

res <- ahmatrix(mat)
}
\author{
Daryanaz Dargahi
}

