\name{ahmatrixObj-class}
\docType{class}
\alias{ahmatrixObj}
\alias{ahmatrixObj-class}
\alias{ahmatrixObj,ANY-method}
\title{Class \code{ahmatrixObj}}

\description{
An object that stores a pairwise comparison matrix.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ahmatrixObj", ...)}.
}

\section{Slots}{
  \describe{
    \item{\code{ahp_matrix}:}{A pairwise comparison matrix.}
  }
}

\author{
Daryanaz Dargahi <daryanazdargahi@gmail.com>
}
