% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp.R
\name{ahp}
\alias{ahp}
\title{ahp}
\usage{
ahp(x, simulation = 500)
}
\arguments{
\item{x}{a pairwise comparison matrix (PCM) with diagonal values equal 1 and a[i,j] = 1/a[j,i].}

\item{simulation}{the simulation size in order to compute Satty's inconsistency. The default value is 500.}
}
\value{
An S4 object including a numeric vector of AHP weights and Satty's inconsistency.
}
\description{
Computing AHP weights as well as Satty's inconsistency.
}
\examples{
mat <- matrix(c(1,2,5, 1/2,1,3, 1/5,1/3,1), nrow = 3, ncol = 3, byrow = TRUE)
res <- ahp(mat, simulation = 500)
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. A scaling method for priorities in hierarchical structures. Journal of Mathematical Psychology, 15(3):234-281, 1977.

T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
}

