% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crplot.R
\name{crplot}
\alias{crplot}
\title{crplot}
\usage{
crplot(srcfile, fontsize = 15, xcex = 10, ycex = 10, angle = 90,
  xlab = "ID", ylab = "ICR", main = NULL)
}
\arguments{
\item{srcfile}{a numeric vector of individual CR}

\item{fontsize}{the font size of plot title, x and y axis labels. The default value is 15.}

\item{xcex,ycex}{the font size of x and y axis, respectively. The default values is 10.}

\item{angle}{the angle of the labels on x axis}

\item{xlab,ylab}{the label to be shown on the x and y axis, respectively.}

\item{main}{the plot title}
}
\value{
An object created by 'ggplot'.
}
\description{
Plotting the consistency ratio (CR) of individual judgements. According to Satty, a pairwise comparison matrix is considered to be consistent if CR is equal or less than 0.1. Therefore, CRs equal to or less than 0.1 are colored in green (Pass) and CRs greater than 0.1 are colored in red (Failed).
}
\examples{
data <- c(0.1132, 0.0142, 0.0324, 0.10075, 0.0883)
names(data) <- c('individual_1','individual_2','individual_3','individual_4','individual_5')
crplot(data, fontsize = 15, xcex = 10, ycex = 10, xlab = 'ID', ylab = 'ICR', main = 'Individuals consistency ratio')
}
\author{
Daryanaz Dargahi
}

