\name{epitope.PCM-data}
\docType{data}
\alias{epitope.PCM}
\alias{epitopePCM}
\alias{epitope.PCM-data}
\title{A pairwise comparison matrix (PCM)}
\description{
epitope.PCM is a criteria with two subcriteria in the prioritization of a set of genes of interest. This file includes a PCM of the two subcriteria including number of epitopes and size of epitopes. The pairwise comparisons are performed according to the Saaty's fundamental scale. 
}
\author{
Daryanaz Dargahi
}



