\name{epitope_num_rating-data}
\docType{data}
\alias{epitope_num_rating}
\alias{epitopeNumRating}
\alias{epitope_num_rating-data}
\title{Rating categories and idealised priorities}
\description{
The rating categories and computed idealised priorities of alternatives (genes) with respect to their number of epitopes. This matrix is computed by the rating() function. 
}
\author{
Daryanaz Dargahi
}
