\name{epitope_size_rating-data}
\docType{data}
\alias{epitope_size_rating}
\alias{epitopeSizeRating}
\alias{epitope_size_rating-data}
\title{Rating categories and idealised priorities}
\description{
The rating categories and computed idealised priorities of alternatives (genes) with respect to their epitope size. This matrix is computed by the rating() function. 
}
\author{
Daryanaz Dargahi
}



