\name{epitopeNum.PCM-data}
\docType{data}
\alias{epitopeNum.PCM}
\alias{epitopeNumPCM}
\alias{epitopeNum.PCM-data}
\title{A pairwise comparison matrix (PCM)}
\description{
EpitopeNum.PCM is a subcriteria (of the Epitope criterion) in the prioritization of a set of genes of interest. This file includes a PCM of the genes with respect to the number of epitopes according to the UniProt annotation (uniprot.org). The pairwise comparisons are performed according to the Saaty's fundamental scale. 
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
}


