\name{freq_exp_rating-data}
\docType{data}
\alias{freq_exp_rating}
\alias{freqExpRating}
\alias{freq_exp_rating-data}
\title{Rating categories and idealised priorities}
\description{
The rating categories and computed idealised priorities of alternatives (genes) with respect to their expression frequency in tumor tissue. This matrix is computed by the rating() function. 
}
\author{
Daryanaz Dargahi
}


