% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregation.R
\name{gaggregate}
\alias{gaggregate}
\title{gaggregate}
\usage{
gaggregate(srcfile, method = "geometric", simulation = 500)
}
\arguments{
\item{srcfile}{a matrix with one or two columns. Column one (required) includes the path (location) to each individual pairwise comparison matrix and column two (optional) includes the individual weights. The matrix rowname is individuals identifier.}

\item{method}{two methods are available for aggregation of individual opinions, (1) arithmetic, which compute the arithmetic mean of individual priorities, (2) geometric, which computes the geometric mean of individual PCMs. If individuals are assigned with a weight, the weighted arithmetic/geometric mean will be computed. The default method is 'geometric'.}

\item{simulation}{simulation size for computation of Satty's inconsistency. The default value is 500.}
}
\value{
An S4 object including group PCM/prioritise;

If geometric mean is used, the returning object includes: aggregated group PCM (AIJ), group consistency ratio (GCR), individual consistency ratios (ICR), consensus index (CI), and priority matrix (IP).

If aritmetic mean is used, the returning object includes: agrregated group priority (AIP), individual consistency ratios (ICR), and priority matrix (IP).
}
\description{
Aggregating individual judgments (pairwise comparison matrices - PCMs) into a group judgement or group priority.
}
\examples{
mat <- matrix(nrow = 4, ncol = 1, data = NA)
mat[,1] <- c(system.file('extdata','ind1.tsv',package = 'Prize'),
            system.file('extdata','ind2.tsv',package = 'Prize'),
            system.file('extdata','ind3.tsv',package = 'Prize'),
            system.file('extdata','ind4.tsv',package = 'Prize'))
rownames(mat) <- c('ind1','ind2','ind3', 'ind4')
colnames(mat) <- c('individual_judgement')

# non-weighted aggregation
res <- gaggregate(srcfile = mat, method = 'geometric', simulation = 500)

# weighted aggregation
# Decision makers are assigned with a priority value based on their specialization and perspectives.
mat <- cbind(mat, c(0.35,0.25,0.15,0.25))
colnames(mat)[2] <- 'individual_weight'

res <- gaggregate(srcfile = mat, method = 'geometric', simulation = 500)
}
\author{
Daryanaz Dargahi
}
\references{
E. Forman and K. Peniwati. Aggregating individual judgments and priorities with the analytic hierarchy process. European Journal of Operational Research, 108(1):165-169, 1998.
}

