\name{genes.list-data}
\docType{data}
\alias{genes.list}
\alias{genesList}
\alias{genes.list-data}
\title{A pairwise comparison matrix (PCM)}
\description{
A matrix of genes to prioritize based on their tumor and normal expression, frequency of expression in tumor tissue, and number and size of epitopes.
}
\author{
Daryanaz Dargahi
}



