\name{ind4-data}
\docType{data}
\alias{ind4}
\alias{ind4-data}
\title{An individual judgement matrix}
\description{
A numeric square matrix generated based on Saaty's fundamental scale. This matrix includes judgements of a decision maker.
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
}



