\name{normal_exp_rating-data}
\docType{data}
\alias{normal_exp_rating}
\alias{normalExpRating}
\alias{normal_exp_rating-data}
\title{Rating categories and idealised priorities}
\description{
The rating categories and computed idealised priorities of alternatives (genes) with respect to their expression in healthy (normal) tissues. This matrix is computed by the rating() function. 
}
\author{
Daryanaz Dargahi
}

