\name{number-data}
\docType{data}
\alias{number}
\alias{number-data}
\title{A pairwise comparison matrix (PCM)}
\description{
Number is a PCM of the categories defined for the number of epitopes subcriteria. The categories include: Single and Multiple. The pairwise comparisons are performed according to the Saaty's fundamental scale. 
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
}





