\name{numEpitope_alternative_category-data}
\docType{data}
\alias{numEpitope_alternative_category}
\alias{numEpitopeAlternativeCategory}
\alias{numEpitope_alternative_category-data}
\title{A pairwise comparison matrix (PCM)}
\description{
A matrix assigning each alternative (gene) to a category with respect to their number of epitopes. The categories include: Single and Multiple.
}
\author{
Daryanaz Dargahi
}


