% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{pipeline}
\usage{
pipeline(srcfile, model, simulation = 500)
}
\arguments{
\item{srcfile}{a character matrix, where the first column specifies the hierarchy order, the second column includes elements IDs, and the third column includes the path to the PCM/priority matrices (See the example below).}

\item{model}{the AHP computation model. Choose from relative and rating models. If using the relative model pairwise comparison matrices must be provided for the evaluation of alternatives. However, if using the raing model, rating matrices must be provided for the evaluation of alternatives.}

\item{simulation}{simulation size for computation of Saaty's inconsistency}
}
\value{
An S4 object including the ahp wight and consistancy measures, and data structures to visualize with ahplot(), rainbowplot(), and wplot().
}
\description{
a pipeline for prioritization estimation using analytic hierarchy process (AHP), which supports both relative and rating AHP models.
}
\examples{
mat <- matrix(nrow = 7, ncol = 3, data = NA)
mat[,1] <- c('0', '1','2','3','4','4.1','4.2')
mat[,2] <- c('Prioritization_of_DE_genes','Tumor_expression','Normal_expression',
             'Frequency', 'Epitopes', 'Number_of_epitopes', 'Size_of_epitopes')
mat[,3] <- c(system.file('extdata','aggreg.judgement.tsv',package = 'Prize'),
             system.file('extdata','tumor.PCM.tsv',package = 'Prize'),
             system.file('extdata','normal.PCM.tsv',package = 'Prize'),
             system.file('extdata','freq.PCM.tsv',package = 'Prize'),
             system.file('extdata','epitope.PCM.tsv',package = 'Prize'),
             system.file('extdata','epitopeNum.PCM.tsv',package = 'Prize'),
             system.file('extdata','epitopeLength.PCM.tsv',package = 'Prize'))

result <- pipeline(mat, model = 'relative', simulation = 500)
}
\author{
Daryanaz Dargahi
}

