% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rainbowplot.R
\name{rainbowplot}
\alias{rainbowplot}
\title{rainbowplot}
\usage{
rainbowplot(srcfile, range = NULL, fontsize = 10, xcex = 4, ycex = 8,
  color = "rainbow", xlab = "Total priority score", ylab = "Alternative",
  digit = 3, dist = 0.02, main = NULL)
}
\arguments{
\item{srcfile}{a numeric matrix with alternatives on the rows and criteria/subcriteria on the columns. alternative's AHP weights for each criteria, and the sum of each row on the last column (see example below).}

\item{range}{specifies which alternatives to plot. Either provide a range (e.g. 1:5) or a vector of numbers (e.g. c(2,5,12,20)). if range is NULL then all alternatives will be plotted. The default value is NULL.}

\item{fontsize}{the font size of the plot title, x and y axis labels, and legend. The default value is 10.}

\item{xcex}{the font size of the labels on the bars. The default values is 4.}

\item{ycex}{the font size of the y axis. The default values is 8.}

\item{color}{the color palette to fill bars. Either provide a vector of n colors, where n is the number of criteria/subcriteria, or choose from the following palettes 'rainbow, heat, terrain, topo, cm'. The default value is rainbow.}

\item{xlab,ylab}{the label of the x and y axis, respectively.}

\item{digit}{the number of digits after decimal point to be shown on the bars. The default value is 3.}

\item{dist}{the distance between the bar and its lable. The default value is 0.02.}

\item{main}{the plot title}
}
\value{
An object created by 'ggplot'.
}
\description{
Plotting prioritized alternatives in a color-coded barplot, where a color is assigned to each criteria/subcriteria.
}
\examples{
mat <- matrix(c(0.007,0.289,0.033,0.118,0.447, 0.015,0.155,0.015,0.088,0.275, 0.048,0.078,0.007,0.044,0.177, 0.039,0.042,0.003,0.017,0.101),
                nrow = 4, ncol = 5, byrow = TRUE, dimnames = list(c('CA9','MUC16','CD70','MUC1'), c('Tumor_expression','Normal_expression','Style','Frequency','Epitopes')))
rainbowplot(mat, range = NULL, xlab = 'Total priority score' , ylab = 'Alternative', dist = 0.04)
}
\author{
Daryanaz Dargahi
}

