% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rating.R
\name{rating}
\alias{rating}
\title{rating}
\usage{
rating(scale, alternative, NA_category = NULL, simulation = 500)
}
\arguments{
\item{scale}{a pairwise comparison matrix (PCM) of rating categories.}

\item{alternative}{a N by 2 character matrix, where N is the number of alternatives. The matrix includes alternatives on column #1 and the rating category they belong to on column #2.}

\item{NA_category}{a character string or vector which specifies categories with the value of zero. Since zero is not achievable by PCM matrix.}

\item{simulation}{simulation size for computation of Saaty's inconsistency}
}
\value{
An S4 object including the raw and normalized ahp priorities, Satty's inconsistency, and rating matrix.
}
\description{
Estimates idealised priorities of alternatives (the rating AHP model).
}
\examples{
mat <- matrix(nrow = 4, ncol = 4, data = NA)

# The category PCM matrix
rownames(mat) <- c('excellent','good','fair','poor')
colnames(mat) <- c('excellent','good','fair','poor')
mat[1,] <- c(1,2,4,6)
mat[2,] <- c(NA,1,2,4)
mat[3,] <- c(NA,NA,1,2)
mat[4,] <- c(NA,NA,NA,1)

# The alternative matrix
alt <- matrix(nrow = 5, ncol = 2, data = NA)
alt[,1] <- c("Andy", "Emily", "Nina", "Alex", "Jack")
alt[,2] <- c("good", "poor", "good", "fair", "excellent")

result <- rating(mat, alt, simulation = 500)

# Specifying a category with value of zero
alt <- rbind(alt, c('shannon', 'Not_available'))

result <- rating(mat, alt, NA_category = 'Not_available', simulation = 500)
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. Rank from comparisons and from ratings in the analytic hierarchy/network processes. European Journal of Operational Research, 168(2):557-570, January 2006.

T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Resource Allocation. McGraw-Hill, New york, 1980.
}

