\name{tumor.PCM-data}
\docType{data}
\alias{tumor.PCM}
\alias{tumorPCM}
\alias{tumor.PCM-data}
\title{A pairwise comparison matrix (PCM)}
\description{
Tumor.PCM is a criteria (with no subcriteria) in the prioritization of a set of genes of interest. This file includes a PCM of the genes with respect to their expression level in tumor tissue. The pairwise comparisons are performed according to the Saaty's fundamental scale. 
}
\author{
Daryanaz Dargahi
}
\references{
T.L. Saaty. The Analytic Hierarchy Process, Planning, Piority Setting, Re- source Allocation. McGraw-Hill, New york, 1980.
}

