\name{tumor_exp_rating-data}
\docType{data}
\alias{tumor_exp_rating}
\alias{tumorExpRating}
\alias{tumor_exp_rating-data}
\title{Rating categories and idealised priorities}
\description{
The rating categories and computed idealised priorities of alternatives (genes) with respect to their expression in tumor tissue. This matrix is computed by the rating() function. 
}
\author{
Daryanaz Dargahi
}
